+++ core/src/test/java/com/graphhopper/routing/util/WheelchairFlagEncoderTest.java
@@ -274,7 +274,7 @@ public class WheelchairFlagEncoderTest {
     public void testPier() {
         ReaderWay way = new ReaderWay(1);
         way.setTag("man_made", "pier");
-        IntsRef flags = wheelchairEncoder.handleWayTags(encodingManager.createEdgeFlags(), way, wheelchairEncoder.getAccess(way));
+        IntsRef flags = wheelchairEncoder.handleWayTags(encodingManager.createEdgeFlags(), way);
         assertFalse(flags.isEmpty());
     }
 
@@ -282,16 +282,16 @@ public class WheelchairFlagEncoderTest {
     public void testMixSpeedAndSafe() {
         ReaderWay way = new ReaderWay(1);
         way.setTag("highway", "motorway");
-        IntsRef flags = wheelchairEncoder.handleWayTags(encodingManager.createEdgeFlags(), way, wheelchairEncoder.getAccess(way));
+        IntsRef flags = wheelchairEncoder.handleWayTags(encodingManager.createEdgeFlags(), way);
         assertTrue(flags.isEmpty());
 
         way.setTag("sidewalk", "yes");
-        flags = wheelchairEncoder.handleWayTags(encodingManager.createEdgeFlags(), way, wheelchairEncoder.getAccess(way));
+        flags = wheelchairEncoder.handleWayTags(encodingManager.createEdgeFlags(), way);
         assertEquals(5, wheelchairAvSpeedEnc.getDecimal(false, flags), .1);
 
         way.clearTags();
         way.setTag("highway", "track");
-        flags = wheelchairEncoder.handleWayTags(encodingManager.createEdgeFlags(), way, wheelchairEncoder.getAccess(way));
+        flags = wheelchairEncoder.handleWayTags(encodingManager.createEdgeFlags(), way);
         assertEquals(0, wheelchairAvSpeedEnc.getDecimal(false, flags), .1);
     }